requirements.txt
Flask==2.0.3
sqlalchemy==1.3.22
flask-restful==0.3.8
Flask-SQLAlchemy==2.4.4
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.1
tabula-py==2.7.0
